tests/wpt/tests/html/semantics/selectors/pseudo-classes/invalid-after-clone.html

<!doctype html>
<meta charset="utf-8">
<link rel="author" title="Emilio Cobos Ãlvarez" href="mailto:emilio@crisal.io">
<link rel="author" title="Mozilla" href="https://mozilla.org">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/testdriver.js"></script>
<script src="/resources/testdriver-vendor.js"></script>
<input>
<textarea></textarea>
<script>
promise_test(async () => {
  for (let tag of ["input", "textarea"]) {
    let element = document.querySelector(tag);
    await test_driver.send_keys(element, 'something');

    assert_true(element.validity.valid, tag + ' should be valid');

    element.maxLength = 0;
    assert_true(element.matches(":invalid"), tag + ' should match :invalid');
    assert_false(element.validity.valid, tag + ' should be invalid');

    let clone = element.cloneNode(true);
    assert_true(clone.matches(":invalid"), tag + ' clone should match :invalid');
    assert_false(clone.validity.valid, tag + 'clone should be invalid');
  }
}, 'Cloned invalid inputs / textareas with interactive changes get their validity state copied correctly');
</script>
